# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FitCareer API",
        "description": "FitCareer REST API documentation.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API version 1"
        }
    ],
    "paths": {
...
        "/auth/email/verification-notification": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend email verification notification",
                "operationId": "14803dc54ad9c033ed11be3f5a6f2156",
                "responses": {
                    "200": {
                        "description": "Verification notification sent"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify email address",
                "operationId": "26879d184285975dbf699a620b517b30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verified"
                    },
                    "403": {
                        "description": "Invalid verification link"
                    }
                }
            }
        },
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke the current Sanctum access token",
                "operationId": "e1d43531b73d491a174eb34d9e132137",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/candidate/skills": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "List candidate skills",
                "operationId": "4af2e250bf48cb6790646e6ada968c2a",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Attach skill to candidate profile",
                "operationId": "fa85c9fa76907d01a59e10be85c5d2df",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/skills": {
            "get": {
                "tags": [
                    "Candidate"
                ],
                "summary": "Search skills catalog for autocomplete",
                "operationId": "899906b590d300b8cc77fed8fded3d92",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skills returned"
                    }
                }
            }
        },
...
        "/company/profile": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get authenticated company profile",
                "operationId": "ee2cbfdbcfd57212cfff55005bd9fe08",
                "responses": {
                    "200": {
                        "description": "Company profile returned"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Update authenticated company profile",
                "operationId": "faf0e34ba64a702ead89342870d6f803",
                "responses": {
                    "200": {
                        "description": "Company profile updated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/company/profile/logo": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Upload company logo",
                "operationId": "a8e2497688c5abf776300abf5556f883",
                "responses": {
                    "200": {
                        "description": "Logo uploaded"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Delete company logo",
                "operationId": "78433f6331aa2ae77499072af7a6fb49",
                "responses": {
                    "200": {
                        "description": "Logo deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}